# apps/api/docs/handoff/ups-expresosapp-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/schedules": {
      "post": {
        "operationId": "SchedulesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Schedule created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new schedule",
        "tags": [
          "Admin Schedules"
        ]
      },
      "get": {
        "operationId": "SchedulesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "routeId",
            "required": false,
            "in": "query",
            "description": "Filter by route ID",
            "schema": {
              "format": "uuid",
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          },
          {
            "name": "dayOfWeek",
            "required": false,
            "in": "query",
            "description": "Filter by day of week",
            "schema": {
              "example": "MONDAY",
              "type": "string",
              "enum": [
                "MONDAY",
                "TUESDAY",
                "WEDNESDAY",
                "THURSDAY",
                "FRIDAY",
                "SATURDAY",
                "SUNDAY"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of schedules",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedulePaginatedResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List schedules with pagination and filters",
        "tags": [
          "Admin Schedules"
        ]
      }
    },
    "/admin/schedules/{id}": {
      "get": {
        "operationId": "SchedulesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Schedule ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Schedule details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Schedule not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get schedule details",
        "tags": [
          "Admin Schedules"
        ]
      },
      "patch": {
        "operationId": "SchedulesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Schedule ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Schedule updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Schedule not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a schedule",
        "tags": [
          "Admin Schedules"
        ]
      },
      "delete": {
        "operationId": "SchedulesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Schedule ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Schedule deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Schedule not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a schedule",
        "tags": [
          "Admin Schedules"
        ]
      }
    },
...
  },
  "info": {
    "title": "UPS ExpresosApp API",
    "description": "Backend API for UPS ExpresosApp MVP - Institutional transport management system for Universidad Politécnica Salesiana. This phase includes: OTP-based authentication, route management, stops, schedules, vehicles, drivers, notices, and audit logging.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateScheduleDto": {
        "type": "object",
        "properties": {
          "routeId": {
            "type": "string",
            "format": "uuid",
            "description": "Route ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "dayOfWeek": {
            "type": "string",
            "description": "Day of week",
            "enum": [
              "MONDAY",
              "TUESDAY",
              "WEDNESDAY",
              "THURSDAY",
              "FRIDAY",
              "SATURDAY",
              "SUNDAY"
            ],
            "example": "MONDAY"
          },
          "direction": {
            "type": "string",
            "description": "Direction",
            "example": "Norte"
          },
          "departureTime": {
            "type": "string",
            "description": "Departure time in HH:mm format",
            "example": "07:30"
          },
          "approximateArrivalTime": {
            "type": "string",
            "description": "Approximate arrival time in HH:mm format",
            "example": "08:15"
          },
          "status": {
            "type": "string",
            "description": "Schedule status",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "default": "ACTIVE"
          }
        },
        "required": [
          "routeId",
          "dayOfWeek",
          "direction",
          "departureTime"
        ]
      },
      "ScheduleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "routeId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "MONDAY",
              "TUESDAY",
              "WEDNESDAY",
              "THURSDAY",
              "FRIDAY",
              "SATURDAY",
              "SUNDAY"
            ],
            "example": "MONDAY"
          },
          "direction": {
            "type": "string",
            "example": "Norte"
          },
          "departureTime": {
            "type": "string",
            "example": "07:30"
          },
          "approximateArrivalTime": {
            "type": "string",
            "nullable": true,
            "description": "Approximate arrival time",
            "example": "08:15"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "example": "ACTIVE"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "routeId",
          "dayOfWeek",
          "direction",
          "departureTime",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "SchedulePaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleResponseDto"
            }
          },
          "meta": {
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateScheduleDto": {
        "type": "object",
        "properties": {
          "routeId": {
            "type": "string",
            "format": "uuid",
            "description": "Route ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "dayOfWeek": {
            "type": "string",
            "description": "Day of week",
            "enum": [
              "MONDAY",
              "TUESDAY",
              "WEDNESDAY",
              "THURSDAY",
              "FRIDAY",
              "SATURDAY",
              "SUNDAY"
            ],
            "example": "MONDAY"
          },
          "direction": {
            "type": "string",
            "description": "Direction",
            "example": "Norte"
          },
          "departureTime": {
            "type": "string",
            "description": "Departure time in HH:mm format",
            "example": "07:30"
          },
          "approximateArrivalTime": {
            "type": "string",
            "description": "Approximate arrival time in HH:mm format",
            "example": "08:15"
          },
          "status": {
            "type": "string",
            "description": "Schedule status",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "default": "ACTIVE"
          }
        }
      },
...
    }
  }
}